fixed bugs

+++ app/Helpers/Helper.php
@@ -135,9 +135,13 @@ if (!function_exists('deleteFile')) {
 	function deleteFile($upload_id)
 	{
 		$upload = Uploads::find($upload_id);
-		Storage::disk('public')->delete($upload->file_path);
-		$upload->delete();
-		return true;
+		if($upload){
+			if($upload->file_path){
+				Storage::disk('public')->delete($upload->file_path);
+			}
+			$upload->delete();
+			return true;
+		}
 	}
 }
